Move back to separate ruby class file
