+++ lib/seek/filtering/boolean_filter.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+module Seek
+  module Filtering
+    class BooleanFilter < Filter
+      def options(collection, active_values = [])
+        collection = apply_joins(collection) || []
+        true_result = apply(collection, ['true'])
+        false_result = apply(collection, ['false'])
+        true_active_actions = active_values.include? 'true'
+        false_active_actions = active_values.include? 'false'
+
+        false_label = label_mapping ? label_mapping&.call(false)&.first : 'false'
+        true_label = label_mapping ? label_mapping&.call(true)&.first : 'true'
+
+        [
+          Seek::Filtering::Option.new(true_label, 'true', true_result.count, true_active_actions, { replace_filters: true }),
+          Seek::Filtering::Option.new(false_label, 'false', false_result.count, false_active_actions, { replace_filters: true })
+        ]
+      end
+    end
+  end
+end
